+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.14.0
 
 - generated code uses absolute imports
+- account for change in Dream in which `Virtual` decorator requires OpenAPI shape
+- throw a more detailed error when DreamSerializer.attribute used to render a non-database, non-Virtual decorated property
 
 ## 1.13.0